+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.20.0
  - feat: adding different frames to `NesDialog`s
+ - feat: allow `NesFileExplorer` to have custom icons
 
 # 0.19.0
  - feat: adding `NesSectionHeader`